# docs/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: LamTo Resident API
  version: v1
paths:
...
  /api/v1/registration-requests:
    post:
      operationId: registration_create
      tags:
      - registration
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RegistrationCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RegistrationCreateRequest'
        required: true
      security:
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationSubmission'
          description: ''
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation failed (code=validation_failed).
        '409':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: A request conflicts with existing state.
        '429':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Too many authentication attempts (code=throttled).
  /api/v1/registration-requests/status:
    get:
      operationId: registration_status
      parameters:
      - in: header
        name: X-Registration-Status-Token
        schema:
          type: string
        required: true
      tags:
      - registration
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationStatus'
          description: ''
        '400':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation failed (code=validation_failed).
        '404':
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Resource not found (code=not_found).
  /api/v1/registration/options:
    get:
      operationId: registration_options
      tags:
      - registration
      security:
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegistrationBuilding'
          description: ''
...
components:
  schemas:
...
    Problem:
      type: object
      description: |-
        RFC 9457 problem+json body with a LamTo stable machine ``code`` (spec 3.1).

        ``detail`` is developer English; the Flutter client owns Vietnamese UI copy
        keyed off ``code``. ``errors`` is present only for validation failures.
      properties:
        type:
          type: string
          description: Problem type URI reference; usually about:blank.
        title:
          type: string
          description: Short human-readable summary (HTTP status phrase).
        status:
          type: integer
          description: HTTP status code.
        code:
          type: string
          description: Stable machine code for client branching (e.g. validation_failed,
            authentication_failed, not_authenticated, permission_denied, not_found,
            occupancy_selection_required, throttled).
        detail:
          type: string
          description: Developer-English explanation; not shown raw to residents.
        errors:
          type: object
          additionalProperties: {}
          description: Per-field validation errors when code is validation_failed.
            Values are lists of {message, code} objects (may nest for non-field errors).
      required:
      - code
      - status
      - title
      - type
...
    RegistrationBuilding:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        units:
          type: array
          items:
            $ref: '#/components/schemas/RegistrationUnit'
      required:
      - id
      - name
      - units
    RegistrationCreateRequest:
      type: object
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 150
        phone:
          type: string
          minLength: 1
          maxLength: 32
        email:
          type: string
        password:
          type: string
          writeOnly: true
          minLength: 1
        building_id:
          type: integer
        unit_id:
          type: integer
      required:
      - building_id
      - full_name
      - password
      - phone
      - unit_id
    RegistrationStatus:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/RegistrationStatusEnum'
        phone:
          type: string
        building:
          type: string
        unit:
          type: string
        rejection_reason:
          type: string
      required:
      - building
      - phone
      - status
      - unit
    RegistrationStatusEnum:
      enum:
      - PENDING
      - APPROVED
      - REJECTED
      - EXPIRED
      type: string
      description: |-
        * `PENDING` - Đang chờ
        * `APPROVED` - Đã phê duyệt
        * `REJECTED` - Đã từ chối
        * `EXPIRED` - Hết hạn
    RegistrationSubmission:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/RegistrationStatusEnum'
        status_token:
          type: string
        phone:
          type: string
      required:
      - phone
      - status
      - status_token
    RegistrationUnit:
      type: object
      properties:
        id:
          type: integer
        label:
          type: string
      required:
      - id
      - label
...
  securitySchemes:
    GateDevice:
      type: apiKey
      in: header
      name: Authorization
      description: GateDevice <credential>
    knoxApiToken:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"